Added background task demo to DummyAdapter
Demonstrates how to use a thread pool executor to run a background
task in an adapter for e.g. background monitoring of a subsystem
from within an adapter.

Note that, while this mechanism has the advantage that long-running
background tasks will not block the main IOLoop, it may require
the use of synchronisation primitives to guard access to underlying
data and/or resources. The alternative would be to use the tornado
PeriodicCallback mechanism to perform monitoring within the IOLoop
context.