Refactored default route handling to be configurable.
DefaultRoute now explicitly maps onto a static handler, rather than
simply redirecting to a path in the static_path defined globally. The
static path is now a configuration option, either resolved relative to the
module source or as an absolute path.

This mechanism is much cleaner in terms of static URLs, allowing a top-level
directory-like URL to be served from static content.

Closes #10