Refactored LogCaptureFilter to deal with handlers.
Avoided need for repeated boilerplate code in test classes using
LogCaptureFilter to add filter to a handler in the root logger. Also
tests if root logger has any handlers and, if not, adds a MemoryHandler
so that messages are processed but not emitted. This avoids breaking the
log capture behaviour of nosetests while allowing the capture filter to
work if nose log capture is disabled in instances where a module is being
tested that has no handler defined.