Moved odin server package up to top-level of repo.
Beginning tests for separating core odin server from project-specific plugins. In order to test
installing server as depedency in plugin, the server setup.py needs to be at the top-level of
the repo so that it can be run when pulling in the server package from github as a dependency link.
This branch is a test to verify if that works.