Make proxy adapter tests work with tornado 6.
The background proxy test server has been refactored to work correctly
in tornado 6 (and earlier versions). Initialisaiton of the web
application and server now takes place in the explicitly initialsied
IOLoop context of the background thread running the test server. This
allows the server to be stopped cleanly under all allowed combinations
of python 2/3 and tornado 5/6.