Make ParameterTree leaf node responses symmetric with and without metadata.
This change makes the parameter tree treatment of leaf nodes symmetric with and
without metadata, i.e. a get() call to a leaf node will return a dict with a
single 'value' entry. The behaviour for higher subtrees is also modified so that
the last level of the subtree path is not repeated for the key of the response.

The set() behaviour is also improved, allowing single leaf node values to be set
with a payload that is a dict with a single 'value' entry.