Refactor the ApiAdapter class to support the adapter-controller pattern.
This commit refactors the ApiAdapter class to support the adapter-controller pattern while
retaining backward-compatiblity with adapters that implement their own HTTP verb handler methods.
The adapter-controller pattern allows a derived adapter to simply define a controller class and
error class and exploit the default behaviour of ApiAdapter to handle all requests and call
the appropriate methods in the controller. The BaseController provides an abstract base class
that defines the interface a derived controller must implement.

Also refactored the structure of support classes and decorator methods, along with associated
test classes and cases.