Make ParameterTree leaf node responses symmetric (#73)
* Make ParameterTree leaf node responses symmetric with and without metadata.

This change makes the parameter tree treatment of leaf nodes symmetric with and
without metadata, i.e. a get() call to a leaf node will return a dict with a
single 'value' entry. The behaviour for higher subtrees is also modified so that
the last level of the subtree path is not repeated for the key of the response.

The set() behaviour is also improved, allowing single leaf node values to be set
with a payload that is a dict with a single 'value' entry.

* Remove reserved tree-level metadata fields.

This commit removes the reserved tree-level metadata fields 'name'
and 'description', allowing those to be used as parameter keys and
preventing issues with explicit gets.

* Remove stray print() statements and replace commented asserts in tests

* Remove redundant level check when populating tree

Closes #71
Closes #58